I think the original simulator is fixed.

+++ src/c/bpopsim/cPopulation.h
@@ -41,7 +41,7 @@ namespace bpopsim {
     uint32_t m_total_mutations;           //Number of mutations so far
     uint32_t m_total_subpopulations_lost; //Number of subpopulations that no longer exist
     uint16_t m_seed;                      //Random number generator
-    uint32_t m_genotype_count;            //used to assign node ids in tree, should equal number of nodes
+    uint64_t m_genotype_count;            //used to assign node ids in tree, should equal number of nodes
     uint64_t m_muller_rez;                //Vertical Resolution of muller plot... set at command line (default: 2500)
     
     // @JEB: An uint16_t rather than a uint because this can go negative by a few cells
